+++ hyperdbg/hyperkd/code/debugger/core/Debugger.c
@@ -359,7 +359,7 @@ DebuggerUninitialize()
     //
     // Uninitialize the HyperTrace (if it was initialized)
     //
-    HyperTraceUnInit();
+    HyperTraceUninit();
 
     //
     // Free the Pool manager