+++ hyperdbg/libhyperdbg/code/debugger/commands/debugging-commands/exit.cpp
@@ -17,6 +17,8 @@
 extern HANDLE  g_DeviceHandle;
 extern BOOLEAN g_IsConnectedToHyperDbgLocally;
 extern BOOLEAN g_IsSerialConnectedToRemoteDebuggee;
+extern BOOLEAN g_IsKdModuleLoaded;
+extern BOOLEAN g_IsVmmModuleLoaded;
 
 /**
  * @brief help of the exit command
@@ -58,19 +60,14 @@ CommandExit(vector<CommandToken> CommandTokens, string Command)
         //
 
         //
-        // unload and exit if the vmm module is loaded
+        // unload and exit if the any module is loaded
         //
         if (g_DeviceHandle)
         {
             //
-            // Unload the VMM module first because it is using the driver and then unload the KD module
+            // Unload all modules
             //
-            HyperDbgUnloadVmm();
-
-            //
-            // Unload the KD module
-            //
-            HyperDbgUnloadKd();
+            HyperDbgUnloadAllModules();
         }
     }
     else if (g_IsSerialConnectedToRemoteDebuggee)